Refactor code for import script
