+++ app/views/reimbursements/_item_breakdown.html.erb
@@ -42,7 +42,7 @@
       
       <div class="d-flex flex-column mx-auto my-2">
         <p class="text-center fw-bold me-2 mb-0">Nothing to display yet.</p>
-        <%= link_to new_reimbursement_path, class: "btn btn-sm btn-primary rounded-pill" do %>
+        <%= link_to new_reimbursement_path(employee_id: params[:employee_id]), class: "btn btn-sm btn-primary rounded-pill" do %>
           File Reimbursement <i class="bi bi-arrow-right-circle"></i>
         <% end %>
       </div>
