+++ config/routes.rb
@@ -14,6 +14,4 @@ Rails.application.routes.draw do
   # Defines the root path route ("/")
   root "reimbursements#index"
   resources :reimbursements, only: [ :index, :new, :create ]
-
-  post "admin/reimbursements/import_data", to: "reimbursements#import_data", as: "admin_import_reimbursement_data"
 end
